pyproject.toml
[tool.poetry]
name = "version-checker"
version = "0.0.3"
description = ""
authors = ["Cédric Bonhomme <cedric@cedricbonhomme.org>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "*"
Flask = "*"
Flask-SQLAlchemy = "*"
Flask-Login = "*"
Flask-Principal = "*"
Flask-WTF = "*"
WTForms = "*"
flask-migrate = "*"
svgwrite = "*"
flask-csv = "*"
pycryptodome = "*"
flask-paginate = "*"
psycopg2-binary = "*"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pre-commit = "^2.20.0"
pyupgrade = "^2.37.3"
reorder-python-imports = "^3.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
